pyproject.toml
[project]
name = "capstone-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=6.1.0",
    "docling>=2.86.0",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langchain-postgres>=0.0.17",
    "langgraph>=1.1.6",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
]
